+++ packages/input-manager/src/integrations/skaarhoj/index.ts
@@ -172,7 +172,7 @@ export class SkaarhojDevice extends Device {
 			if (((feedback.tally ?? 0) & Tally.NEXT) !== 0) {
 				tallyColor = 3
 			}
-			if (((feedback.tally ?? 0) & Tally.CURRENT) !== 0) {
+			if (((feedback.tally ?? 0) & Tally.ACTIVE) !== 0) {
 				tallyColor = 2
 			}
 		} else {