+++ packages/input-manager/src/integrations/xkeys/index.ts
@@ -177,8 +177,8 @@ export class XKeysDevice extends Device {
 
 	private static selectKeyBacklight(tally: Tally | undefined, classNames: string[] | undefined): string | null {
 		if (classNames === undefined) return null
-		if (tally !== undefined && (tally & Tally.CURRENT) === Tally.CURRENT) return Colors.RED
-		if (tally !== undefined && (tally & Tally.NEXT) === Tally.CURRENT) return Colors.GREEN
+		if (tally !== undefined && (tally & Tally.ACTIVE) === Tally.ACTIVE) return Colors.RED
+		if (tally !== undefined && (tally & Tally.NEXT) === Tally.ACTIVE) return Colors.GREEN
 		if (classNames.includes(ClassNames.AD_LIB)) return Colors.ORANGE
 		return null
 	}