+++ CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/Sofie-Automation/sofie-input-gateway/compare/v0.3.1-alpha.3...v0.4.0) (2025-07-01)
+
+
+### Features
+
+* docker image ([02313be](https://github.com/Sofie-Automation/sofie-input-gateway/commit/02313be8338174228487a0da00a756472e22df6c))
+* support ip streamdecks ([#26](https://github.com/Sofie-Automation/sofie-input-gateway/issues/26)) ([1ed0913](https://github.com/Sofie-Automation/sofie-input-gateway/commit/1ed091345cd61288a41c59717196e3def2bb7a00))
+
+
+
+
+
+## [0.3.1-alpha.3](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.2...v0.3.1-alpha.3) (2025-03-24)
+
+
+### Bug Fixes
+
+* add suffix to release ZIP ([0d931b0](https://github.com/nrkno/sofie-input-gateway/commit/0d931b0400eb7650574e7fe13d64bdd8d97a17aa))
+
+
+
+
+
+## [0.3.1-alpha.2](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.1...v0.3.1-alpha.2) (2025-03-24)
+
+**Note:** Version bump only for package input-gateway-packages
+
+
+
+
+
+## [0.3.1-alpha.1](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.0...v0.3.1-alpha.1) (2025-01-28)
+
+
+### Features
+
+* add more debug logging for tracing rendering issues with skia-canvas ([ab06124](https://github.com/nrkno/sofie-input-gateway/commit/ab061245bf5e92578b400e5d91d1af28bc8443e3))
+
+
+
+
+
+## [0.3.1-alpha.0](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.0...v0.3.1-alpha.0) (2025-01-24)
+
+
+### Bug Fixes
+
+* resolve build issue related to p-queue ([8e850f8](https://github.com/nrkno/sofie-input-gateway/commit/8e850f84bf883862f41c78f3992b459c3ec93adb))
+* **Stream Deck:** update elgato-stream-deck library to resolve an issue with slow discovery of HID devices ([fbcb790](https://github.com/nrkno/sofie-input-gateway/commit/fbcb790c42be28b277aa8298f6432bb6271a5d6b))
+
+
+
+
+
 # [0.3.0](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.0-alpha.2...v0.3.0) (2024-10-01)
 
 **Note:** Version bump only for package input-gateway-packages