+++ packages/input-manager/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.3...v0.4.0) (2025-07-01)
+
+
+### Features
+
+* support ip streamdecks ([#26](https://github.com/nrkno/sofie-input-gateway/issues/26)) ([1ed0913](https://github.com/nrkno/sofie-input-gateway/commit/1ed091345cd61288a41c59717196e3def2bb7a00))
+
+
+
+
+
+## [0.3.1-alpha.3](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.2...v0.3.1-alpha.3) (2025-03-24)
+
+**Note:** Version bump only for package @sofie-automation/input-manager
+
+
+
+
+
+## [0.3.1-alpha.2](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.1...v0.3.1-alpha.2) (2025-03-24)
+
+**Note:** Version bump only for package @sofie-automation/input-manager
+
+
+
+
+
+## [0.3.1-alpha.1](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.1-alpha.0...v0.3.1-alpha.1) (2025-01-28)
+
+
+### Features
+
+* add more debug logging for tracing rendering issues with skia-canvas ([ab06124](https://github.com/nrkno/sofie-input-gateway/commit/ab061245bf5e92578b400e5d91d1af28bc8443e3))
+
+
+
+
+
+## [0.3.1-alpha.0](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.0...v0.3.1-alpha.0) (2025-01-24)
+
+
+### Bug Fixes
+
+* **Stream Deck:** update elgato-stream-deck library to resolve an issue with slow discovery of HID devices ([fbcb790](https://github.com/nrkno/sofie-input-gateway/commit/fbcb790c42be28b277aa8298f6432bb6271a5d6b))
+
+
+
+
+
 # [0.3.0](https://github.com/nrkno/sofie-input-gateway/compare/v0.3.0-alpha.2...v0.3.0) (2024-10-01)
 
 **Note:** Version bump only for package @sofie-automation/input-manager